feat: show purchase option on trial expired page
If the user's subscription enters a cancelled state, they cannot re-add
a subscription via the customer portal. Instead, if a user enters this
state, they'll need to go through a new checkout process to purchase a
new subscription.

To keep it simple, we'll just provide both options on this page, and let
the human do the thinking.